reconstruct_main_function
