+++ cmd/helmfile/deps.go
@@ -1,0 +1,28 @@
+package helmfile
+
+import (
+	"github.com/helmfile/helmfile/pkg/app"
+	"github.com/helmfile/helmfile/pkg/config"
+	"github.com/urfave/cli"
+)
+
+func addDepsSubcommand(cliApp *cli.App) {
+	cliApp.Commands = append(cliApp.Commands, cli.Command{
+		Name:  "deps",
+		Usage: "update charts based on their requirements",
+		Flags: []cli.Flag{
+			cli.StringFlag{
+				Name:  "args",
+				Value: "",
+				Usage: "pass args to helm exec",
+			},
+			cli.BoolFlag{
+				Name:  "skip-repos",
+				Usage: `skip running "helm repo update" before running "helm dependency build"`,
+			},
+		},
+		Action: Action(func(a *app.App, c config.ConfigImpl) error {
+			return a.Deps(c)
+		}),
+	})
+}
